Update results submission
